refactor: extract neo4j_helpers.py from neo4j_service.py
Move standalone helpers (_REVERSE_DIRECTION, _location_context_query,
_LOCATION_CONTEXT_QUERIES, _gen_id, _node_to_location, _node_to_npc,
_node_to_item, _dispatch_change) to neo4j_helpers.py (321 lines).

neo4j_service.py: 970 → 684 lines (-286, -29.5%). Neo4jWorldService
class remains at 645 lines — cohesive unit not split further.
